requirements.txt
langchain
openai
requests
pydantic
python-dotenv